+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - fix the wrongly implemented getOptions for [LeanRunConfiguration.getOptions]
 - remove a dependent log in CHANGELOG.md in build.gradle.kts
-- replace the debouncing of goto workspace symbols/classes to triggering suffix
+- add suffix triggering or debounce triggering strategies for "goto workspace symbols/classes" in settings
 - make "open in find tool window" of "goto workspace symbols/classes" work
 
 ## [0.1.9] - 2024-12-16